+++ src/components/App.css
@@ -4,16 +4,7 @@
   margin-left: 2%;
 }
 
-.left-section {
-  width: 20%;
-}
-
 .right-section {
   right: 0;
   width: 74%;
-}
-
-.lang-btn {
-  font-size: 12px;
-  margin: 0px 5px;
 }
