+++ src/components/header/Navigation.css
@@ -1,0 +1,4 @@
+.lang-btn {
+  font-size: 12px;
+  margin: 0px 5px;
+}
