+++ src/components/sidebar/AuthorThumb.css
