+++ src/components/sidebar/AuthorThumb.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import './Author.css'
+import './AuthorThumb.css'
 
-const Author = ({ authorName, authorImage, setAuthor }) => (
+const AuthorThumb = ({ authorName, authorImage, setAuthor }) => (
   <div className="author" onClick={() => setAuthor(authorName)}>
     {authorImage ? <img src={authorImage} alt={authorName} /> : ''}
     {authorName}
   </div>
 )
 
-export default Author
+export default AuthorThumb
