+++ src/components/sidebar/Authors.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import Author from './Author'
+import AuthorThumb from './AuthorThumb'
 
 const Authors = ({ authorImages, authors, setAuthor }) => {
   const sortedAuthors = [...authors].sort()
   const preparedAuthors = sortedAuthors.map((author, i) =>
-    <Author key={i} authorName={author} authorImage={authorImages.get(author)} setAuthor={setAuthor} />
+    <AuthorThumb key={i} authorName={author} authorImage={authorImages.get(author)} setAuthor={setAuthor} />
   )
   return (
     <div className="authors">
