+++ src/components/sidebar/Sidebar.jsx
@@ -4,7 +4,7 @@ import Authors from './Authors'
 import './Sidebar.css'
 
 const Sidebar = ({ authorImages, authors, setPhrase, setAuthor, filterAuthors }) => (
-  <aside>
+  <aside className="sidebar">
     <div className="sidebar-inner">
       <Filters setPhrase={setPhrase} setAuthor={setAuthor} filterAuthors={filterAuthors} />
       <Authors authorImages={authorImages} authors={authors} setAuthor={setAuthor} />
